+++ test/test/schema-encoder.ts
@@ -115,6 +115,54 @@ describe('SchemaEncoder', () => {
             value: []
           }
         ]
+      },
+      {
+        schema:
+          'string contractGroupName,string ipfsHash,string readmeIpfsHash,address[] contractAddresses,string[] contractNames,string[] contractNetworks,string[] contractIpfsHashes',
+        decodedSchema: [
+          {
+            name: 'contractGroupName',
+            type: 'string',
+            signature: 'string contractGroupName',
+            value: ''
+          },
+          {
+            name: 'ipfsHash',
+            type: 'string',
+            signature: 'string ipfsHash',
+            value: ''
+          },
+          {
+            name: 'readmeIpfsHash',
+            type: 'string',
+            signature: 'string readmeIpfsHash',
+            value: ''
+          },
+          {
+            name: 'contractAddresses',
+            type: 'address[]',
+            signature: 'address[] contractAddresses',
+            value: []
+          },
+          {
+            name: 'contractNames',
+            type: 'string[]',
+            signature: 'string[] contractNames',
+            value: []
+          },
+          {
+            name: 'contractNetworks',
+            type: 'string[]',
+            signature: 'string[] contractNetworks',
+            value: []
+          },
+          {
+            name: 'contractIpfsHashes',
+            type: 'string[]',
+            signature: 'string[] contractIpfsHashes',
+            value: []
+          }
+        ]
       }
     ]) {
       context(schema, () => {
@@ -122,18 +170,9 @@ describe('SchemaEncoder', () => {
           const schemaEncoder = new SchemaEncoder(schema);
           expect(schemaEncoder.schema).to.deep.equal(decodedSchema);
         });
-      });
-    }
 
-    for (const schema of [
-      'boo like',
-      'adss contractAddress,bool trusted',
-      'bytes32 eventId,uint8 ticketType,uint10000 ticketNum',
-      'bytes32 eventId,uint8 ticketType,ticketNum'
-    ]) {
-      context(schema, () => {
-        it('should throw', () => {
-          expect(() => new SchemaEncoder(schema)).to.throw(Error);
+        it('should verify', () => {
+          expect(SchemaEncoder.isSchemaValid(schema)).to.be.true;
         });
       });
     }
@@ -675,6 +714,21 @@ describe('SchemaEncoder', () => {
           });
         });
       }
+    }
+  });
+
+  describe('verification', () => {
+    for (const schema of [
+      'boo like',
+      'adss contractAddress,bool trusted',
+      'bytes32 eventId,uint8 ticketType,uint10000 ticketNum',
+      'bytes32 eventId,uint8 ticketType,ticketNum'
+    ]) {
+      context(schema, () => {
+        it('should unable to verify', () => {
+          expect(SchemaEncoder.isSchemaValid(schema)).to.be.false;
+        });
+      });
     }
   });