+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.5.0
+
+- Add schema verification function
+- Fix handling of schemas with "ipfsHash" variables
+
 ## 2.4.1
 
 - Export Private Data types